raidentweaks: semantic versioning fix
It has been detected that all but one version has semanic versioning applied incorrectly, this commit will fix that and future releases will ship with semanic versioning correctly applied.

Below are the old and new versions. The old ones are on the left and the new ones on the right.

- v4.6.6 = v2.2.0
- v4.5.5 = v2.1.0
- v4.5.4 (HOTFIX) = v2.0.0
- v4.5.4 = v1.7.0
- v3.4.4 (HOTFIX) = v1.6.1
- v3.4.4 = v1.6.0
- v3.3.3 = v1.5.0
- v3.2.0 = v1.4.0
- v3.1.1 = v1.3.0
- v2.0.0 = v1.2.0
- v1.9.8 = v1.1.0
- v1.0.0 = v1.0.0

Signed-off-by: Raiden Ishigami <contact.raidenishi69@gmail.com>