raidentweaks: Update license for removed APKs and add them back [1/2]
This commit replaces the license for two APKs that were previously removed from the project. The previous license was causing an 'INSTALL_FAILED_VERIFICATION_FAILURE' error during the installation process. The new license is more suitable for our use case and should prevent this error from occurring again. The APKs are now being added back to the project, ready for further development."

Signed-off-by: Raiden Ishigami <contact.raidenishi69@gmail.com>