raidentweaks: add signal handler for Ctrl+C interruption
Implemented a signal handler using the 'trap' command in the build.sh to capture the SIGINT signal (Ctrl+C interruption). When the signal is received, the script now executes a function to send a message to Telegram indicating that the execution was interrupted. This enhances user experience by providing feedback when the script is interrupted.

Signed-off-by: Raiden Ishigami <contact.raidenishi69@gmail.com>